backend/requirements.txt
# Django
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Auth
PyJWT==2.8.0
cryptography==41.0.7

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utils
python-dateutil==2.8.2
requests==2.31.0

# Environment
python-decouple==3.8
dj-database-url==2.1.0
